Update patch set 2
Patch Set 2:

I am not convinced that we need this.  If we do, I think we have a bigger problem too.

The ril in cm-12.1 did not have a GET_RADIO_CAPABILITIES command.  Therefore, you wouldn't really expect our blobs to necessarily have that command.  But, it apparently does?

I instrumented the java code which uses this resource and haven't seen a log message generated from it in about a week.

I modified the ril to remove the new commands and suddenly I started seeing the messages from the instrumented code.

I haven't looked to verify that it's getting the right values.

But, I'm thinking that this is probably not required and that we really do have an implementation of that new command.

Patch-set: 2