Update patch set 2
Patch Set 2:

I haven't been using this patch because I'm not sure whether or not we need it.  I generally have working data but often need to toggle airplane mode to get rid of the "!" on the LTE (captive portal check?  Other problem?).

I'm going to start running with this patch to see if it makes it any better.

Patch-set: 2