Update patch set 1
Patch Set 1:

David, which part of that commit?  The part that writes the pids?  Probably to use cpusets right.

At the moment though, I was trying to fix the fact that our device was acting as if all the processes were single threaded.  There was very little multi-core usage (although more than I would expect, given that all the sets were limited to cpu 0).

Patch-set: 1