Update patch set 1
Patch Set 1:

(see also my previous comment on this change):

Looking more carefully with this change reverted I see that every time it exits deep sleep it is taking on average 11ms to disable the CPUs and 39ms to wake them back up.  That is 50ms to do anything that transitions from deep sleep to wakeup back to deep sleep.

That said, it takes around 200ms to enable/disable all the devices to do a resume / suspend cycle.

I see around the same total time to do the enable / disable cycle with and without this commit.

Without this commit, it takes more time to wakeup the cpus but then uses multiple cores for the suspend/resume of the devices.

With this commit we save the time to wakeup the cpus but then take more elapsed time to suspend the devices.

I have no idea if this will really save any battery and whether or not it will create any problems with random background media playback, etc.

Patch-set: 1