Update patch set 2
Patch Set 2: Code-Review-1

In general, it's actually pretty cool.  I'm pleasantly surprised.

* Video recording defaults to QCIF resolution which is, umm, unfortunate
* Lack of touch focus for picture taking is sad.

Given that there are a multitude of other options, if we can get video recording to default to the right resolution then I'm good with merging this.

It would definitely be nice to get the focus going though.  I'll miss that one myself.

Patch-set: 2