Update patch set 5
Patch Set 5:

Good point to actually remove it...

We will need to remove thermald from here as well.  I believe (but haven't looked at it in a very long time) that thermald is only used in recovery so I haven't touched it.

But I do want to remove it before this becomes official because we shouldn't be blobs in the device tree...

Patch-set: 5