Update patch set 4
Patch Set 4: Code-Review+1

FYI, relabeling /factory/userdata_footer does not work in the case where it does not exist already. In that case it fails still wanting "create" perms for vold there as well. Give vold create perms and skip the relabel. This works as advertised to create a new footer for the first time.

Patch-set: 4