Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

I'm going to kill this for now.  If we're lucky, it actually won't make much of a difference in battery utilization when we are idle.  If we aren't lucky, we'll have to see if we can come up with a nice backport of upstream changes that fix it.

I tried pulling commits like we talked about, but I wasn't feeling very confident about it.

Patch-set: 1
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+2