Update patch set 3
Patch Set 3: Code-Review-2

I agree with Jason.  This needs a prebuilt x86 image be compiled and uploaded to maven each time there is a new release.  Until someone is motivated enough to do that and change the tooling to pull the right version, this is a bad idea.

If it does ever work now (I swear it worked for me at one time) it will be running in Houdini and this is one application that really doesn't want to run in emulation mode!

Patch-set: 3
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-2